models/requirements.txt
pillow
tensorflow
numpy
# other dependencies...